feat(web): Moved debug button to the top of the factory next to the buttons. Moved debug flag into the app store so it can be used in multiple places
